+++ ChangeLog.md
@@ -3,6 +3,8 @@ Web Authentication change log
 
 ## ?.?.? / ????-??-??
 
+* **Heads up**: Refactor `web.auth.Flow` to an abstract base class
+  (@thekid)
 * **Heads up**: Refactored `Authentication` to be a base class of the
   two implementations, `web.auth.Basic` and `web.auth.SessionBased`
   (@thekid)